requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.42.0
django-cleanup==5.1.0
django-crispy-forms==1.9.2
django-htmlmin==0.11.0
django-optimized-image==0.3.0
django-phonenumber-field==5.0.0
html5lib==1.1
idna==2.10
oauthlib==3.1.0
phonenumbers==8.12.11
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tinify==1.5.1
urllib3==1.25.10
webencodings==0.5.1
